requirements.txt
# Core Dependencies
torch>=1.9.0
transformers>=4.21.0
sentence-transformers==2.2.0  # Pinned to specific version
huggingface-hub==0.17.3  # Pinned to compatible version

# Web Framework
flask==2.0.1
fastapi==0.105.0
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.0.1
pandas==2.1.3
openpyxl==3.1.2
beautifulsoup4==4.12.2

# Machine Learning
numpy==1.26.2
scikit-learn==1.3.2

# Utilities
aiofiles==23.2.1
python-magic==0.4.27
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2
pytz==2023.3.post1